pyproject.toml
[project]
name = "servio"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bleach>=6.3.0",
    "django-allauth[socialaccount]>=65.12.1",
    "django-autoslug>=1.9.9",
    "django-extensions>=4.1",
    "django-htmx>=1.27.0",
    "mysqlclient>=2.2.7",
    "nanoid>=2.0.0",
    "pillow>=12.0.0",
    "pydantic>=2.12.5",
    "pymysql>=1.1.2",
    "pyopenssl>=25.3.0",
    "python-dateutil>=2.9.0.post0",
    "python-decouple>=3.8",
    "uuid6>=2025.0.1",
    "werkzeug>=3.1.4",
    "whitenoise>=6.11.0",
]
